+++ .eslintrc.js
@@ -34,6 +34,9 @@ module.exports = {
       { blankLine: "any", prev: ["const", "let", "var"], next: ["const", "let", "var"] },
     ],
     "eslint/no-var-requires": 0,
+    "react/prop-types": 0,
+    "import/namespace": 0,
+    "import/no-unresolved": "off",
   },
   env: {
     browser: true,
