+++ components/accountSetting/BioUpdateBox/style.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const ProfileBioInput = styled.textarea`
+  width: 34.375rem;
+  height: 8.125rem;
+  margin-left: 0.375rem;
+  margin-right: 2.625rem;
+  outline: none;
+  border: none;
+  font-family: "Pretendard-Regular";
+  font-size: 0.875rem;
+  color: #999893;
+  resize: none;
+  background-color: #f5f5f7;
+  border-radius: 0.5rem;
+  padding: 0.875rem 1rem;
+`;
