+++ components/accountSetting/NicknameUpdateBox/style.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const ErrorMessage = styled.span`
+  font-family: "Pretendard-Medium";
+  font-size: 0.75rem;
+  color: #b6b6b6;
+  margin-top: 0.2rem;
+  margin-left: 0.2rem;
+`;
